First push
Here it is. An initial push of the game. Rough wireframes and all. Currently it has 7 levels. Incoming will be talents and the first boss! Not to mention the obvious missing textures and sounds! Also working on an improved class for level creation to reduce physical typing needed for level creation.